UI tweaks
Remove the fairly useless (and unreliable) open in browser (it rarely
opens in anything useful).

Add tooltip to Export button.

Change help tooltip to be more reflective of the actual useful (and
functional) interactions.  At this point it is just the Zoom to selection
feature.  Most of the other interactions revert themselves when the
mouse is released, and they are not that useful anyway.